MOD: refector swap asset component
